20230523/back/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
dj-rest-auth==7.0.1
Django==4.2.21
django-allauth==65.8.1
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.16.0
idna==3.6
pillow==11.2.1
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.13.2
tzdata==2023.4
urllib3==2.2.1
